invoice-processor-durable/requirements.txt
azure-functions
azure-functions-durable
requests
openai
python-dotenv
azure-ai-projects
azure-ai-agents
azure-identity
